Add `set` in `knob::State`.
This change is nessecary because simply changing `normal_param.value` is
not sufficient to set the knob to the value we want it to have. This is
because `continuous_normal` ends up overwriting the value we want to set
when handling an event, so the correct thing to do is to change both
`normal_param.value` and `continuous_normal`. A similar change should be
made for all the other widgets.