core/Normal: replace `From<f32>` with `TryFrom<f32>`
... and rename `Normal::new` and `Normal::set`.

The implementations of `Normal::new` and (the `From<f32>` trait) clips
the provided float to the normal range. While this ensures the `Normal`
is valid, it prevents the user from handling out of range errors
whithout checking the value first and then calling the clipping
constructor or setter.

This commit adds the `TryFrom<f32>` implementation which fails if the
provided value is out of range.

Previous behaviour is still applicable in most cases, but the clipping
should be made explicite IMO. I removed the `From<f32>` implementation
which would conflict with `TryFrom<f32>` anyway and renamed
`Normal::new` as `Normal::from_clipped`.

For the same reasons, I also renamed `Normal::set` as
`Normal::set_clipped` and added the faillible `Normal::try_set`.
`Normal::set_clipped` is now used wherever I found it to be useful in
the code base.

In order to avoid the need to explicitely `use std::convert::TryFrom`,
I lifted the crate's edition to "2021". Note that `iced` uses edition
"2021" anyway and the minimum Rust supported version is at least
1.65.0 since iced uses GATs in at least one place [1].

Added unit tests for the renamed and added functions.

[1]: https://github.com/iced-rs/iced/blob/f38e7fcac2e3505d11577ade1757a77ca2a544ea/graphics/src/image/storage.rs#L12