Fix value discontinuity resulting from an external change
When a value discontinuity occured as a result of an external change (as opposed
to a change resulting from a user interaction with the widget), a subsequent
wheel scroll would start from the value prior to the discontinuity. This was
due to the `continous_normal` not being synchronized at discontinuity.

This didn't occur when dragging the widget because the `continuous_normal` is
synchronized when the mouse button is pressed.

This commit solves the issue by keeping previous normal value and by
synchronizing `continuous_normal` when an elligible discontinuity is detected.
Synchronization is also factorized instead of being scattered at the various
use cases sites.

All widgets are updated except `XYPad`, which doesn't use the wheel scroll
event. This widget could have been updated by factorizing the
`continuous_normal` synchronization, but that would add the `prev_normal` field
to the `State`, which is not needed. I decided not to add the extra field and
keep current implementation untouched.